* Fixed a nasty bug --- in toc.cyc, we were copying the bound expression for comprehensions of the form:
  new {for i < e : 0}

Here, e is copied once to support the allocation, and once to serve
as the loop bound.  The reason for this is simply that the "new"
part is translated independent of the comprehension --- it's a long
standing bug which was commented in toc, but it had a bad effect
which caused a core dump in the tile benchmark.  

The problem is that we weren't doing a deep copy of e, so we ended
up with the same abstract syntax being shared.  In turn, this was
problematic when e included a statement expression that had a
declaration of a temp within it.  The reason this was a problem is
that the temp-renaming code imperatively updates the code to do
the renaming.  Urgh.  We didn't trip across this elsewhere, simply
because there weren't any comprehensions with effects in the 
bounds...

I started to work-around this by using deep_copy_exp, but this
had two problems.  The first is that deep_copy_exp couldn't handle
statement expressions.  I extended it so that it could (sort of),
but didn't bother to get the nested declarations that get copied
rewired properly.  So be forewarned that deep_copy_exp will 
not work properly with statements embedded in it...  Still, we
should be using deep_copy_exp throughout the back-end to avoid
sharing problems, so I modified toc.cyc, remove_aggregates, and
lower to use it instead of the shallow copy.  

The second problem is that of course, the bounds expression could
have side effects, so duplicating it was not a good idea anyway.
So I fixed this in toc.cyc within the translation of New_e.  

Once again, I am cursing the fact that we didn't do translation
in a functional manner :-)

This should fix the core dump problem we had with the tile benchmark.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4924 581db07f-6e62-4d40-9733-d432c4eebb07