* Added a distinct kind of coercion (Subset_coercion) to absyn.h which is why so many of the files changed.
* Modified treatment of coercions to subset types so that a run-time
check is inserted.  More specifically, if you cast an expression e
to a subset type @subset(T x | assn[x]) either explicitly or implicitly,
we rewrite the cast to be of the form ({let t = e; assert(assn[temp]); t}).
The vcgen infrastructure attempts to optimize away the assert where
possible and issues a warning (if -warnassert or -Wall) is turned on.
This is very similar to what happens with @checks clauses on functions.

Note that this effectively prevents subset types from being initialized
at top-level (or for static variables).  

* Added built-in exception "Assert" to the top-level environment.  This
is the exception that's thrown when an assert(-) fails.  Note that
the built-in assert is different from the one provided by <assert.h>
(which should really be deprecated in favor of the built-in mechanism.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4944 581db07f-6e62-4d40-9733-d432c4eebb07