fixing the bug that Mike found out for the following code: void test(int i) {     int arr[4];     arr[i] = 1;     *(arr + i + 3) = 2; } int main() {   test(2); }
what happened is that during vcgen, the statement
arr[i] = 1; will generate a term offseti(arr,i,type) to
represent the value of the left hand side of the assignment, and
type was given as the type of arr, which is a thin pointer with
size 4.  This is wrong, because arr+i no long has size 4, but
size 4-i. later we have arr+i+3, and since we thought that
arr+i has size 4, it thinks that arr+i+3 is safe.

I made a quick fix to it by changing the type to NULL for now.
But we really should change it to a type with some info but the
right info.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@5086 581db07f-6e62-4d40-9733-d432c4eebb07